saving progress

+++ app/src/main/java/com/example/screensnap/presentation/Video.kt
@@ -6,6 +6,6 @@ import android.net.Uri
 data class Video(
     val uri: Uri,
     val name: String,
-    val duration: Int,
-    val size: Int
+    val duration: Long,
+    val size: Long,
 )
